Minor change with the apiErrorMessages, added a message/msg option
This was done to resolve a tiny issue with the following:
aem.send res, "200", error: "Request was cancelled"

This was not really an error, but the old helper used "error" to replace the message.

It has been changed to the following, which is clearer in its objective.
aem.send res, "200", msg: "Request was cancelled"