So much totally fucked shit (I'm tired, this is crazy, I just gave up, so fix me maybe)
It would appear that the select field is comparing against the ng-model as a string...
Since the campaign ads and the actual full ad list ads don't have the same structure,
I stripped them of all keys but name + id + status.

Yet it still doesn't work. I had debug statements all over the place, they are IDENTICAL,
even comparing them manually in the controller on load and printing to the console finds
the campaign ad in the full ad list.

So much wtf. Going to sleep. Speed, please try and fix this :D Otherwise I do it 'morrow.